requirements.txt
# Steam
requests

# Epic
legendary-gl

# Gog
# (uses system SQLite)

# Web interface
fastapi
uvicorn[standard]
python-multipart
jinja2

# Desktop application
pywebview
pystray  # Optional: system tray icon support

# IGDB integration
python-dotenv

# Metacritic scraping
beautifulsoup4
# Testing
pytest